requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
librosa>=0.9.0
soundfile>=0.10.0
umap-learn>=0.5.0
plotly>=5.0.0
transformers>=4.20.0
nltk>=3.7
jupyter>=1.0.0
ipykernel>=6.0.0